requirements.txt
aws-cdk
aws-cdk-lib>=2.0.0
constructs>=10.0.0
boto3>=1.18.0
requests>=2.25.0  
pandas>=1.1.0     
numpy>=1.19.0     
pytest>=6.0.0     
moto>=2.0.0       
black>=21.0.0     
flake8>=3.8.0 
python-dotenv>=1.0.1   
 
